Test notes for search are now created via paste
