+++ SimplenoteUITests/SimplenoteUITestsSearch.swift
@@ -47,7 +47,7 @@ class SimplenoteUISmokeTestsSearch: XCTestCase {
         Sidebar.tagsDeleteAll()
         NoteList.openAllNotes()
 
-        allNotes.forEach { NoteList.createNoteThenLeaveEditor($0) }
+        allNotes.forEach { NoteList.createNoteThenLeaveEditor($0, viaPaste: true) }
 }
 
     override func setUpWithError() throws {
